\name{QuadRoot}
\alias{QuadRoot}
\title{Quadratic Root for any Quadratic Equation}
\usage{
QuadRoot(inputdata)
}
\description{
It is used to find simple quadratic roots from any quadratic equation
}

\arguments{
  \item{inputdata}{
coefficient vector of quadratic equation
}
}

\value{It returns quadratic roots of the equation}

\author{Pankaj Das}

\examples{
dataset=c(2,5,-3)
QuadRoot(dataset)
}
\keyword{Quadractic Root}
